requirements.txt
asgiref==3.9.2
Django==5.2.6
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-modeltranslation==0.19.17
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
